tracker_bot/requirements.txt
# Include shared dependencies
-r ../requirements.txt

# Telegram bot functionality
python-telegram-bot==20.7

# Google Sheets integration
gspread==5.12.4
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1